requirements.txt
telethon==1.34.0
aiosqlite==0.20.0
PyGithub==2.3.0
PySocks==1.7.1
pytest==8.2.2
pytest-asyncio==0.23.7